refactoring survey options to be on creation modal
