+++ ui-admin/src/study/consents/CreateConsentModal.tsx
@@ -17,11 +17,19 @@ export default function CreateConsentModal({ studyEnvContext, onDismiss }: {
     studyEnvContext: StudyEnvContextT
     onDismiss: () => void }) {
   const [isLoading, setIsLoading] = useState(false)
-  const { formName, formStableId, clearFields, nameInput, stableIdInput } = useFormCreationNameFields()
+  const [form, setForm] = useState<ConsentForm>({
+    createdAt: 0,
+    id: '',
+    lastUpdatedAt: 0,
+    version: 1,
+    content: '{"pages":[]}',
+    name: '',
+    stableId: ''
+  })
+  const { clearFields, NameInput, StableIdInput } = useFormCreationNameFields(form, setForm)
 
   const portalContext = useContext(PortalContext) as PortalContextT
   const navigate = useNavigate()
-
 
   const attachConsentToEnv = async (createdConsent: ConsentForm) => {
     await Api.createConfiguredConsent(studyEnvContext.portal.shortcode,
@@ -45,18 +53,14 @@ export default function CreateConsentModal({ studyEnvContext, onDismiss }: {
 
   const createConsent = async () => {
     await doApiLoad(async () => {
-      const createdConsent = await Api.createNewConsentForm(studyEnvContext.portal.shortcode,
-        {
-          createdAt: 0, id: '', lastUpdatedAt: 0, version: 1,
-          content: '{"pages":[]}', name: formName, stableId: formStableId
-        }
-      )
+      const createdConsent = await Api.createNewConsentForm(studyEnvContext.portal.shortcode, form)
+
       Store.addNotification(successNotification('Consent form created'))
 
       await attachConsentToEnv(createdConsent)
 
       await portalContext.reloadPortal(studyEnvContext.portal.shortcode)
-      navigate(`consentForms/${formStableId}`)
+      navigate(`consentForms/${form.stableId}`)
       setIsLoading(false)
       onDismiss()
     }, { setIsLoading })
@@ -72,18 +76,18 @@ export default function CreateConsentModal({ studyEnvContext, onDismiss }: {
     <Modal.Body>
       <form onSubmit={e => e.preventDefault()}>
         <label className="form-label" htmlFor="inputFormName">Consent Name</label>
-        { nameInput}
+        { NameInput}
         <label className="form-label mt-3" htmlFor="inputFormStableId">Consent Stable ID</label>
         <InfoPopup content={`A stable and unique identifier for the consent. 
                                       May be shown in exported datasets.`}/>
-        { stableIdInput }
+        { StableIdInput }
       </form>
     </Modal.Body>
     <Modal.Footer>
       <LoadingSpinner isLoading={isLoading}>
         <button
           className="btn btn-primary"
-          disabled={!formName || !formStableId}
+          disabled={!form.name || !form.stableId}
           onClick={createConsent}
         >Create</button>
         <button className="btn btn-secondary" onClick={() => {
