+++ ui-admin/src/study/surveys/CreatePreEnrollSurveyModal.tsx
@@ -2,7 +2,7 @@ import React, { useContext, useState } from 'react'
 import { StudyEnvContextT } from '../StudyEnvironmentRouter'
 import { PortalContext, PortalContextT } from 'portal/PortalProvider'
 import { useNavigate } from 'react-router-dom'
-import Api from 'api/api'
+import Api, { Survey } from 'api/api'
 import { Store } from 'react-notifications-component'
 import { successNotification } from 'util/notifications'
 import Modal from 'react-bootstrap/Modal'
@@ -19,15 +19,21 @@ export default function CreatePreEnrollSurveyModal({ studyEnvContext, onDismiss 
 
   const portalContext = useContext(PortalContext) as PortalContextT
   const navigate = useNavigate()
-  const { formName, formStableId, clearFields, nameInput, stableIdInput } = useFormCreationNameFields()
+  const [form, setForm] = useState<Survey>({
+    ...defaultSurvey,
+    stableId: '',
+    name: '',
+    surveyType: 'RESEARCH',
+    version: 1,
+    content: '{"pages":[]}',
+    id: '',
+    createdAt: new Date().getDate(),
+    lastUpdatedAt: new Date().getDate()
+  })
+  const { clearFields, NameInput, StableIdInput } = useFormCreationNameFields(form, setForm)
   const createSurvey =async () => {
     await doApiLoad(async () => {
-      const createdSurvey = await Api.createNewSurvey(studyEnvContext.portal.shortcode,
-        {
-          ...defaultSurvey,
-          createdAt: 0, id: '', lastUpdatedAt: 0, version: 1, surveyType: 'RESEARCH',
-          content: '{"pages":[]}', name: formName, stableId: formStableId
-        })
+      const createdSurvey = await Api.createNewSurvey(studyEnvContext.portal.shortcode, form)
       Store.addNotification(successNotification('Survey created'))
       try {
         await Api.updateStudyEnvironment(studyEnvContext.portal.shortcode,
@@ -43,7 +49,7 @@ export default function CreatePreEnrollSurveyModal({ studyEnvContext, onDismiss 
       }
 
       await portalContext.reloadPortal(studyEnvContext.portal.shortcode)
-      navigate(`preEnroll/${formStableId}`)
+      navigate(`preEnroll/${form.stableId}`)
     }, { setIsLoading })
     onDismiss()
   }
@@ -59,17 +65,17 @@ export default function CreatePreEnrollSurveyModal({ studyEnvContext, onDismiss 
     <Modal.Body>
       <form onSubmit={e => e.preventDefault()}>
         <label className="form-label" htmlFor="inputFormName">Survey Name</label>
-        { nameInput }
+        { NameInput }
         <label className="form-label mt-3" htmlFor="inputFormStableId">Survey Stable ID</label>
         <InfoPopup content={'A stable and unique identifier for the form. May be shown in exported datasets.'}/>
-        { stableIdInput }
+        { StableIdInput }
       </form>
     </Modal.Body>
     <Modal.Footer>
       <LoadingSpinner isLoading={isLoading}>
         <button
           className="btn btn-primary"
-          disabled={!formName || !formStableId}
+          disabled={!form.name || !form.stableId}
           onClick={createSurvey}
         >Create</button>
         <button className="btn btn-secondary" onClick={() => {
