+++ ui-admin/src/study/surveys/CreateSurveyModal.test.tsx
@@ -101,9 +101,9 @@ describe('CreateSurveyModal', () => {
         assignToExistingEnrollees: true,
         blurb: 'Testing out the marketing blurb...',
         content: expect.stringContaining('{"pages":[{"elements":[{"type":"html","name":"outreach_content_'),
-        createdAt: 0,
+        createdAt: expect.anything(),
+        lastUpdatedAt: expect.anything(),
         id: '',
-        lastUpdatedAt: 0,
         name: 'Test Marketing',
         stableId: 'testMarketing',
         surveyType: 'OUTREACH',
@@ -141,9 +141,9 @@ describe('CreateSurveyModal', () => {
         blurb: 'Testing out the screener blurb...',
         assignToExistingEnrollees: true,
         content: '{"pages":[]}',
-        createdAt: 0,
+        createdAt: expect.anything(),
+        lastUpdatedAt: expect.anything(),
         id: '',
-        lastUpdatedAt: 0,
         name: 'Test Screener',
         stableId: 'testScreener',
         surveyType: 'OUTREACH',
