+++ ui-admin/src/study/surveys/FormOptionsModal.tsx
@@ -1,0 +1,88 @@
+import React from 'react'
+import { Survey, VersionedForm } from 'api/api'
+import Modal from 'react-bootstrap/Modal'
+import { Button } from 'components/forms/Button'
+import { faArrowRight } from '@fortawesome/free-solid-svg-icons'
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
+import { SaveableFormProps } from './SurveyView'
+import { DocsKey, ZendeskLink } from 'util/zendeskUtils'
+import InfoPopup from 'components/forms/InfoPopup'
+
+/** component for selecting versions of a form */
+export default function FormOptionsModal({
+  workingForm, updateWorkingForm, onDismiss
+}:
+                                          { workingForm: VersionedForm,
+                                            updateWorkingForm: (props: SaveableFormProps) => void,
+                                            onDismiss: () => void}) {
+  const isSurvey = !!(workingForm as Survey).surveyType
+
+  return <Modal show={true} onHide={onDismiss} size="lg">
+    <Modal.Header closeButton>
+      <Modal.Title>{workingForm.name} - configuration</Modal.Title>
+    </Modal.Header>
+    <Modal.Body>
+      <FormOptions workingForm={workingForm} updateWorkingForm={updateWorkingForm}/>
+      <div className="fw-light fst-italic mt-4">
+        Note: you must  &quot;Save&quot; the form
+        for changes to these options to take effect.
+      </div>
+    </Modal.Body>
+    <Modal.Footer>
+      <Button variant="secondary" onClick={onDismiss}>
+        Done
+      </Button>
+    </Modal.Footer>
+  </Modal>
+}
+
+/**
+ *
+ */
+export const FormOptions = ({ workingForm, updateWorkingForm }:
+{ workingForm: VersionedForm,
+  updateWorkingForm: (props: SaveableFormProps) => void}) => {
+  const isSurvey = !!(workingForm as Survey).surveyType
+
+  return <>
+    { isSurvey &&
+        <div>
+          <h3 className="h5">Configuration <InfoPopup placement="right" content={<div>
+            See the Options <FontAwesomeIcon icon={faArrowRight}/> Configuration section in
+            our <ZendeskLink doc={DocsKey.SURVEY_EDIT}>survey editing docs</ZendeskLink> for
+            information on the options below.
+          </div>}/></h3>
+          <form className="p-2">
+            <label className="form-label d-block">
+              <input type="checkbox" checked={(workingForm as Survey).required}
+                onChange={e => updateWorkingForm({
+                  ...workingForm, required: e.target.checked
+                })}
+              /> Required
+            </label>
+            <label className="form-label d-block">
+              <input type="checkbox" checked={(workingForm as Survey).assignToAllNewEnrollees}
+                onChange={e => updateWorkingForm({
+                  ...workingForm, assignToAllNewEnrollees: e.target.checked
+                })}
+              /> Auto-assign to new participants
+            </label>
+            <label className="form-label d-block">
+              <input type="checkbox" checked={(workingForm as Survey).assignToExistingEnrollees}
+                onChange={e => updateWorkingForm({
+                  ...workingForm, assignToExistingEnrollees: e.target.checked
+                })}
+              /> Auto-assign to existing participants
+            </label>
+            <label className="form-label d-block">
+              <input type="checkbox" checked={(workingForm as Survey).autoUpdateTaskAssignments}
+                onChange={e => updateWorkingForm({
+                  ...workingForm, autoUpdateTaskAssignments: e.target.checked
+                })}
+              /> Auto-update participant tasks to the latest version of this survey after publishing
+            </label>
+          </form>
+        </div>
+    }
+  </>
+}
